+++ pages/user/profile.tsx
@@ -1,6 +1,7 @@
 import { Container } from "@mui/material";
 import { GetServerSideProps } from "next";
 import { useRouter } from "next/router";
+import { useEffect } from "react";
 import ProfileEditor from "../../components/Profile/ProfileEditor";
 import ProfileRegister from "../../components/Profile/ProfileRegister";
 import { useAuthState } from "../../hook/useAuthState";
@@ -11,7 +12,12 @@ type Props = {
 const ProfilePage = ({ registerMode }: Props) => {
   const router = useRouter();
   const { authState } = useAuthState();
-
+  useEffect(() => {
+    const value = sessionStorage.getItem("register");
+    if (value === null) return;
+    sessionStorage.removeItem("register");
+    router.push("/user/joined");
+  }, []);
   return (
     <>
       {authState.isLoading || !authState.isLogined ? (
